--- a/.gitignore
+++ b/.gitignore
@@ -3,3 +3,6 @@ node_modules
 
 # logs
 npm-debug.log
+
+# built files
+/dist