--- a/index.js
+++ b/index.js
@@ -12,28 +12,35 @@ var y = d * 365.25;
 /**
  * Parse or format the given `val`.
  *
- * Options:
- *
- *  - `long` verbose formatting [false]
- *
  * @param {String|Number} val
- * @param {Object} [options]
+ * @param {Boolean} toLong verbose string formatting [false]
  * @throws {Error} throw an error if val is not a non-empty string or a number
  * @return {String|Number}
  * @api public
  */
 
-export default function(val, options) {
-  options = options || {};
-  var type = typeof val;
+export default function(val, toLong) {
+  var abs,
+    fmt,
+    type = typeof val;
+
   if (type === 'string' && val.length > 0) {
     return parse(val);
-  } else if (type === 'number' && isFinite(val)) {
-    return options.long ? fmtLong(val) : fmtShort(val);
   }
+
+  if (type === 'number' && isFinite(val)) {
+    fmt = plural.bind(plural, toLong, val, (abs = Math.abs(val)));
+
+    if (abs >= d) return fmt(d, 'day', 'd');
+    if (abs >= h) return fmt(h, 'hour', 'h');
+    if (abs >= m) return fmt(m, 'minute', 'm');
+    if (abs >= s) return fmt(s, 'second', 's');
+
+    return val + (toLong ? ' ' : '') + 'ms';
+  }
+
   throw new Error(
-    'val is not a non-empty string or a valid number. val=' +
-      JSON.stringify(val)
+    'val is an empty string or a invalid number. val=' + JSON.stringify(val)
   );
 }
 
@@ -46,7 +53,6 @@ export default function(val, options) {
  */
 
 function parse(str) {
-  str = String(str);
   if (str.length > 100) {
     return;
   }
@@ -98,65 +104,17 @@ function parse(str) {
     case 'ms':
       return n;
     default:
-      return undefined;
+      return;
   }
-}
-
-/**
- * Short format for `ms`.
- *
- * @param {Number} ms
- * @return {String}
- * @api private
- */
-
-function fmtShort(ms) {
-  var msAbs = Math.abs(ms);
-  if (msAbs >= d) {
-    return Math.round(ms / d) + 'd';
-  }
-  if (msAbs >= h) {
-    return Math.round(ms / h) + 'h';
-  }
-  if (msAbs >= m) {
-    return Math.round(ms / m) + 'm';
-  }
-  if (msAbs >= s) {
-    return Math.round(ms / s) + 's';
-  }
-  return ms + 'ms';
-}
-
-/**
- * Long format for `ms`.
- *
- * @param {Number} ms
- * @return {String}
- * @api private
- */
-
-function fmtLong(ms) {
-  var msAbs = Math.abs(ms);
-  if (msAbs >= d) {
-    return plural(ms, msAbs, d, 'day');
-  }
-  if (msAbs >= h) {
-    return plural(ms, msAbs, h, 'hour');
-  }
-  if (msAbs >= m) {
-    return plural(ms, msAbs, m, 'minute');
-  }
-  if (msAbs >= s) {
-    return plural(ms, msAbs, s, 'second');
-  }
-  return ms + ' ms';
 }
 
 /**
  * Pluralization helper.
  */
 
-function plural(ms, msAbs, n, name) {
-  var isPlural = msAbs >= n * 1.5;
-  return Math.round(ms / n) + ' ' + name + (isPlural ? 's' : '');
+function plural(isLong, ms, abs, interval, l1, l2) {
+  return (
+    Math.round(ms / interval) +
+    (isLong ? ' ' + l1 + (abs >= interval * 1.5 ? 's' : '') : l2)
+  );
 }