--- a/package.json
+++ b/package.json
@@ -39,6 +39,7 @@
     "expect.js": "0.3.1",
     "husky": "0.14.3",
     "lint-staged": "5.0.0",
-    "mocha": "4.0.1"
+    "mocha": "4.0.1",
+    "prettier": "1.17.1"
   }
 }