@leerob 2021-03-02
Hey! Thank you for the exploration here. Very interesting!

In the interest of ensuring this package is maintainable, readable, and most importantly stable for the millions who depend on it, we're going to be very selective of the PRs that get merged in.

Closing this as won't fix, but thank you regardless 🙏