@maraisr 2021-03-03
I don't know about that @leerob I'd argue this PR address some pretty serious things—stuff I believe should have been merged.

Few points;
- You mentioned _maintainability_ now having benchmarks means this can be better maintained knowing there is no regression on performance.
- You mentioned readability. Having a variables named `n` is far less readable than that of `num`. But as this is biased, could you work with @lukeed to align some of this concerns?
- Stability, I'd argue this is more stable than whats out there now, as this has tests and benchmark to prove it.

@leerob  I urge you to reconsider this one.