@lukeed 2021-03-03
As mentioned at the top of the PR:

```
Breaking

Changed the 2nd param to be a boolean, toggling long-mode directly
Of course this can be reverted, and I'm open to it, but don't foresee other options :)
```

The breaking change here can easily be reverted, and you maintain the size reduction & performance win.
Even if it weren't, _proper_ versioning allows for there to be breaking changes without affecting existing users.

But I get it – this was just a nice way to express that there's no interest :) no problem