@lukeed 2021-03-03
Gotcha. Well, I can restore the branch & you could reopen the PR, but TBH I'm kinda done waiting & putting effort into this contribution, so you/someone else is going to have to put in the extra bit of elbow grease to resolve conflicts & revert the parameter change. It's been loud & clear that it wasn't/never was wanted, and that has turned me off on investing anything else here.

I'll keep my fork – only because I've already been using it for a year & prefer its API. But again, I'm more than happy for this PR's size & perf improvements to make its way into `ms` because it'll help out more people. My fork will be very insignificant, at best/if anything.

Even though you ultimately 180'd, I appreciate that you took the time to say _something_ @leerob :)