@styfle 2021-03-04
I think the reason this PR was never merged is because its doing too many things.

I would accept a PR doing these things:

- Adds minimal benchmark suite
- Includes prettier directly