@lukeed 2021-03-04
I'm sorry – I generally agree with small PRs, but I really don't think that was the case here:

* there has to be a benchmark to prove the performance improvements
* I had to include/update prettier to actually get the PR in a passing state, since it relied on global installation

At best, the "multiple formats" feature could have been a separate PR, but it was included here to track/prove the size reduction. 

Either way, I would have been happy to address any of this between June 2019-2020 had it been raised. 
I know it's nothing to do with you personally @styfle (or @leerob) – nor is it your fault(s) – but overall this could have been handled much better at any point in time.

TBH I'm still not sure if this (or parts of this) is _actually_ wanted, or if the conversation is only continuing because my fork is public/published, which gave this thread some visibility (sorry for that btw, not intended... just wanted to give proper credit & reference as to _why_ my fork exists). Until I know that answer, I'm not inclined to invest any more time/focus here, but as stated above, I'd still like to make the contribution.

I'm not insisting that you accept the PR by any means. A rejection is perfectly acceptable – even an immediate rejection would have been okay! It's your project & I'll accept whatever the decision may be. It's just that after a long wait period, I'm left confused w/ mixed signals.