@Nytelife26 2021-04-17
I'm not sure how people intend to move this forward, but there is definitely interest. @lukeed, computer science is about science. I'm sorry you had to deal with the whole "if it isn't broken don't fix it" ideology.

If you do not intend to continue working on this, I'm happy to pick it up.