@AryanBeezadhur 2021-12-14
@Nytelife26 he doesn't intend to continue working on it. Instead, he has published his own fork: [lukeed/ms](https://github.com/lukeed/ms).